Completed the search logic
The 3 main searches are now in place with correct logic. These are
searches for; Companies to assoc with, Customers & Suppliers with checks
made on the searches to make sure we don't offer a company twice in the
same place.

Next job is to determine what we do with the results sets and where they
take the user in each case.